requirements.txt
python-telegram-bot==20.7
openai>=1.0.0
python-dotenv
pypdf
python-docx
rapidfuzz
